github.com/twitchscience/spade_edge - Multiple kinesis producers + statsd prefix
Summary:
The edges were not able to keep up. I played with config variables (primarily pushing up the batch size) which helped but was not sufficient, so
this adds the ability to have multiple instances of the producer object (currently the config file is set to 4).

Additionally to make it easier to understand where issues are coming from and to be able to accumulate counters in grafana I modified the stats prefix to

Test Plan: Ran as canary in prod

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3383