Providing more info when logging a large request
Summary:
Instead of enforcing the 500KB limit during the request parsing, we now use the default limit (currently 10mb) but later enforce the limit by checking the content length ourselves. This allows us to collect the data field during the failure in most of the cases and log part of that for debugging.

https://twitchtv.atlassian.net/browse/SCIENG-468

Test Plan: Ran the requests unit test and I'm planning to use the integration env to verify it works on normal flow and also logs when I send a large request.

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: mperrone, #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4107