Add counter for URI over 8 KB
Summary:
The ALBs that we want to use for Edge in the future won't accept requests with
header lines over 8 KB, so we want to move producers submitting data via GET
to start using POST instead.

For now, let's just count the requests which would actually break on the ALBs;
we can log them with the event name (and count the number of GET requests
overall) when we've dealt with Curse and Visage.

Test Plan: Ran in integration and saw the expected count on Grafana.

Reviewers: maasted

Reviewed By: maasted

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5110