Count Edge requests with events in URI
Summary:
Sending events in the URI is a potential failure mode if the events get too
big, so we want to track how often they happen.  As we might have a large
number at first, I added a configurable sampling rate.

I also refactored all checks of the query parameters to use url.Values
objects, which is parsed once when we start handling the request.
Calling r.URL.Query() parses the raw query every time it's called.

Test Plan:
Edge still works in integration.
Created a table in Grafana and verified that events appeared.

Deploy: Will add the sampling rate to the integration and prod config
(1.0 in integration, 0.0001 in prod for now).

Reviewers: ojcm

Reviewed By: ojcm

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5266