Add round-robin option to S3 uploader
Summary:
Performance analysis on spade_edge in production revealed heavy
contention on the input channel for the upload logger in the S3 event
logger.  To ameliorate this, we allow multiple upload loggers in the
event logger and assign events to them in a round-robin fashion.
(Ideally the upload loggers would simply share an input channel, but
this would require changes in the underlying library.)

Test Plan:
Integration test passes.  Edge loadtest on a c4.2xlarge instance
with two upload loggers showed better performance than before the
change, though not quite twice as good as on a c4.xlarge instance;
we should check in production whether there is still contention
somewhere.

Reviewers: lydon

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5771